Fixed routing after reload
